feat(agent): an unproven acceptance criterion blocks the final answer
The contract has known what the work agreed to since #8183 and been able to
answer mid-turn since #8193, and criteria became provable in #8201. Nothing
consulted it. A turn could satisfy every delivery ceremony — todos closed,
complete_step called, verification run — while a criterion the user approved
sat unproven, because those ceremonies check that steps happened, not that
the agreed outcome holds.

finalReadinessCheckFor now folds the contract's outstanding criteria into
the same `missing` list it already builds. One gate, one FinalReadinessError,
one host recovery path: the Goal FSM's auto-continue and the recovery card
work unchanged, which they would not if this were a second parallel refusal.

The existing missingAcceptanceCriteria field asked whether criteria were
*established*; this asks whether they are *proven*, and names which one. A
proof that went stale under a later mutation reads as outstanding again, so
"verified before I changed it" stops counting as verified.

Only a planned turn contributes. Without a plan the contract's requirements
are the todo list, which the checks above already cover, so an unplanned turn
inherits nothing.

Placed before the loop-guard escape deliberately: a criterion the model
cannot prove must still be able to stop asking, or an unprovable acceptance
line ("P95 under 100ms") would loop forever. The guard that already exists
for that keeps working.

Outstanding() now annotates a stale requirement the way it already did a
stale check — "verify it" and "re-verify it, the code moved" are different
instructions and only one is actionable.

Debt moved, not added: readiness left agent.go for final_readiness.go, so
agent.go drops 233 lines (complexity 61->37, function-size 125->103) while
the new file records exactly what moved. Every baseline total falls or holds
— complexity 2099->2096, file-size 110345->110104, function-size 9165->9159,
essay 4085->4082, test-file-size flat.
