+++ internal/agent/agent.go
@@ -1608,221 +1608,11 @@ func (a *Agent) ReadinessResult() ReadinessResult {
 	}
 }
 
-type finalReadinessCheck struct {
-	applies                   bool
-	reason                    string
-	missingProjectChecks      int
-	incompleteTodos           int
-	missingAcceptanceCriteria int
-	missingVerification       int
-	missingReview             int
-	missingSignoff            int
-	missingActionEvidence     int
-	missingMutation           int
-	missingCapabilities       int
-}
-
-func (c finalReadinessCheck) progressSignature() string {
-	return fmt.Sprintf("%d/%d/%d/%d/%d/%d/%d/%d/%d/%d\x00%s",
-		c.missingProjectChecks,
-		c.incompleteTodos,
-		c.missingAcceptanceCriteria,
-		c.missingVerification,
-		c.missingReview,
-		c.missingSignoff,
-		c.missingActionEvidence,
-		c.missingMutation,
-		c.missingCapabilities,
-		boolInt(c.applies),
-		c.reason,
-	)
-}
-
-func (c finalReadinessCheck) missingIDs() []string {
-	missing := make([]string, 0, 9)
-	add := func(id string, count int) {
-		if count > 0 {
-			missing = append(missing, id)
-		}
-	}
-	add("project_check", c.missingProjectChecks)
-	add("todo", c.incompleteTodos)
-	add("criteria", c.missingAcceptanceCriteria)
-	add("verification", c.missingVerification)
-	add("review", c.missingReview)
-	add("signoff", c.missingSignoff)
-	add("action", c.missingActionEvidence)
-	add("mutation", c.missingMutation)
-	add("capability", c.missingCapabilities)
-	return missing
-}
-
 func boolInt(v bool) int {
 	if v {
 		return 1
 	}
 	return 0
-}
-
-func (c finalReadinessCheck) audit(result evidence.ReadinessAuditResult, recovered bool) evidence.ReadinessAudit {
-	return evidence.ReadinessAudit{
-		Result:                    result,
-		Recovered:                 recovered,
-		MissingProjectChecks:      c.missingProjectChecks,
-		IncompleteTodos:           c.incompleteTodos,
-		CommandMismatchMissing:    c.missingProjectChecks,
-		MissingAcceptanceCriteria: c.missingAcceptanceCriteria,
-		MissingVerification:       c.missingVerification,
-		MissingReview:             c.missingReview,
-		MissingSignoff:            c.missingSignoff,
-		MissingActionEvidence:     c.missingActionEvidence,
-		MissingMutation:           c.missingMutation,
-		MissingCapabilities:       c.missingCapabilities,
-	}
-}
-
-func (a *Agent) finalReadinessCheckFor() finalReadinessCheck {
-	if a.evidence == nil || a.ablation.Off(ablation.Evidence) {
-		return finalReadinessCheck{}
-	}
-	var missing []string
-	out := finalReadinessCheck{}
-	// Planning returns a proposal to the controller, which owns the approval gate
-	// and starts a fresh execution turn after Plan is disabled. Delivery completion
-	// requirements, including required capabilities, wait for that execution turn:
-	// forcing them here could make a writer requirement contradict the plan-first
-	// workflow. This is a workflow boundary only; model-initiated tool calls above
-	// still use the normal Permissions/Sandbox path.
-	if a.planMode.Load() {
-		return out
-	}
-	{
-		incomplete, hasTodos := a.evidence.IncompleteLatestTodos()
-		if !hasTodos && a.evidence.HasAnySuccessfulReceipt() {
-			incomplete, hasTodos = a.incompleteCanonicalTodos()
-		}
-		if hasTodos && len(incomplete) > 0 && a.evidence.HasSuccessfulTodoProgressReceipt() {
-			out.applies = true
-			out.incompleteTodos = len(incomplete)
-			missing = append(missing, finalReadinessIncompleteTodos(incomplete))
-		}
-	}
-	writer, hasWriter := a.evidence.LatestSuccessfulWriterIndex()
-	deliveryMutation := false
-	deliveryVerificationOnly := false
-	checkpoint := a.deliveryCheckpoint
-	checkpointApplies := a.deliveryScopeActive && checkpoint.ScopeID == a.deliveryScopeID
-	if a.deliveryProfile {
-		if mutation, ok := a.evidence.LatestSuccessfulMutationIndex(); ok {
-			writer, hasWriter = mutation, true
-			deliveryMutation = true
-		} else if checkpointApplies && checkpoint.PendingMutation {
-			// The mutation happened before a controller rebuild/restart. Treat it as
-			// the baseline so this run can satisfy verification/review/sign-off
-			// without manufacturing another write.
-			writer, hasWriter = -1, true
-			deliveryMutation = true
-		} else if checkpointApplies && checkpoint.MutationObserved {
-			deliveryMutation = true
-		}
-		workObserved := a.evidence.HasSuccessfulWorkReceipt() || (checkpointApplies && checkpoint.WorkObserved)
-		if a.deliveryTaskExpected && !a.deliveryPersistentExpected && !workObserved {
-			out.missingActionEvidence++
-			missing = append(missing, "perform host-observable work for this technical task before answering")
-		}
-		if a.deliveryPersistentExpected && !a.evidence.HasSuccessfulToolReceipt("remember") {
-			out.missingMutation++
-			missing = append(missing, "save the requested durable memory with the remember tool before answering")
-		}
-		if a.deliveryMutationExpected && !deliveryMutation {
-			out.missingMutation++
-			missing = append(missing, "the request requires a state change, but no successful mutation was observed")
-		}
-		if !hasWriter && a.evidence.HasSuccessfulVerificationCommand() {
-			writer, hasWriter = -1, true
-			deliveryVerificationOnly = true
-		}
-		// Required/preferred capability gates apply before the no-writer fast
-		// path below: a user-required Skill/MCP must not be skippable by
-		// answering from ordinary reads alone.
-		if msg := a.capabilityGateFailure(); msg != "" {
-			out.applies = true
-			out.missingCapabilities++
-			missing = append(missing, msg)
-		}
-		if a.deliveryPersistentExpected && !a.deliveryMutationExpected && !a.evidence.HasSuccessfulMutationOtherThan("remember") {
-			// A durable-memory-only request has its own concrete receipt contract.
-			// It must not inherit code-delivery todo/test/diff/review ceremonies;
-			// any unrelated mutation falls through to the full contract below.
-			out.applies = true
-			if len(missing) > 0 {
-				out.reason = strings.Join(missing, "; ")
-			}
-			return out
-		}
-	}
-	if !hasWriter {
-		if len(missing) > 0 {
-			if a.loopGuardAllowsFinal() {
-				return out
-			}
-			out.reason = strings.Join(missing, "; ")
-		}
-		return out
-	}
-	hasProjectChecks := len(a.projectChecks) > 0
-	hasTodoReceipt := a.evidence.HasSuccessfulTodoWrite()
-	if !a.deliveryProfile && !hasProjectChecks && !hasTodoReceipt && len(missing) == 0 {
-		return finalReadinessCheck{}
-	}
-	out.applies = true
-	if a.deliveryProfile {
-		criteriaEstablished := a.deliveryCriteriaEstablished || (checkpointApplies && checkpoint.CriteriaEstablished)
-		if !criteriaEstablished {
-			out.missingAcceptanceCriteria++
-			missing = append(missing, "establish concrete acceptance criteria with todo_write before changing state")
-		}
-		hasCompleteStep := a.evidence.HasSuccessfulCompleteStepAfter(writer)
-		if !hasCompleteStep {
-			out.missingSignoff++
-			missing = append(missing, "call complete_step after the latest mutation")
-		}
-		if !a.evidence.HasSuccessfulDeliverySignoffAfter(writer) {
-			out.missingVerification++
-			missing = append(missing, "run relevant verification after the latest mutation and cite that successful command in complete_step")
-		}
-		if deliveryMutation && !a.evidence.HasSuccessfulReviewAfter(writer) {
-			out.missingReview++
-			missing = append(missing, "inspect the changed result after the latest mutation (read the touched file or run git diff/status)")
-		}
-		if msg := a.deliveryReviewGateFailure(); msg != "" {
-			out.missingReview++
-			missing = append(missing, msg)
-		}
-		// The capability gate already ran before the no-writer fast path above.
-	}
-	for _, check := range a.projectChecks {
-		if deliveryVerificationOnly {
-			break
-		}
-		command := strings.TrimSpace(check.Command)
-		if command == "" {
-			continue
-		}
-		if !a.evidence.HasSuccessfulCommandAfter(command, writer) {
-			out.missingProjectChecks++
-			missing = append(missing, fmt.Sprintf("run %q from %s after the latest write", command, finalReadinessCheckSource(check)))
-		}
-	}
-
-	if len(missing) == 0 {
-		return out
-	}
-	if a.loopGuardAllowsFinal() {
-		return out
-	}
-	out.reason = strings.Join(missing, "; ")
-	return out
 }
 
 // DeliveryCheckpoint returns the compact Goal-scoped delivery state. It is safe
@@ -1891,18 +1681,6 @@ func (a *Agent) deliveryMutationCheckpointReady() bool {
 		a.evidence.HasSuccessfulDeliverySignoffAfter(mutation) &&
 		a.evidence.HasSuccessfulReviewAfter(mutation) &&
 		a.deliveryReviewGateFailure() == ""
-}
-
-func finalReadinessIncompleteTodos(items []evidence.TodoStepMatch) string {
-	parts := make([]string, 0, len(items))
-	for _, item := range items {
-		label := strings.TrimSpace(item.Content)
-		if label == "" {
-			label = fmt.Sprintf("todo %d", item.Index)
-		}
-		parts = append(parts, fmt.Sprintf("%s: %s", label, item.Status))
-	}
-	return "latest successful todo_write still has incomplete items: " + strings.Join(parts, ", ")
 }
 
 func (a *Agent) setTodoState(todos []evidence.TodoItem) {
@@ -2055,17 +1833,6 @@ func toolResultFailed(content string) bool {
 		strings.HasPrefix(content, "blocked:") ||
 		strings.HasPrefix(content, "Error:") ||
 		strings.HasPrefix(content, "[error")
-}
-
-func finalReadinessCheckSource(check instruction.VerifyCheck) string {
-	source := strings.TrimSpace(check.SourcePath)
-	if source == "" {
-		source = "project memory"
-	}
-	if check.Line > 0 {
-		return fmt.Sprintf("%s:%d", source, check.Line)
-	}
-	return source
 }
 
 func shouldNudgeExecutorHandoff(input, answer string) bool {
